full setup of cleaning

+++ GGJ20/Assets/Scripts/Work/Data/Tasks/CleaningTaskScriptableObject.cs
@@ -3,7 +3,7 @@
 [CreateAssetMenu(fileName = "CleaningTask", menuName = "ScriptableObjects/CleaningTask", order = 1)]
 public class CleaningTaskScriptableObject : TaskScriptableObject
 {
-    public int TargetCleanliness;
+    public float TargetCleanliness;
     public override WorkManager.TaskType GetTaskType()
     {
         return WorkManager.TaskType.Cleaning;
